requirements.txt
openai>=1.0.0
requests>=2.25.0
pytest>=6.0.0
python-dotenv>=0.19.0
python-wordpress-xmlrpc>=2.3.0